requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=2.0.0

# RSS and XML parsing
feedparser>=6.0.10

# OpenAI integration
openai>=1.0.0

# Audio processing and transcription
openai-whisper>=20231117

# Optional audio processing (only needed if using local Whisper fallback)
# Uncomment if you want SpeechRecognition fallback:
# SpeechRecognition>=3.10.0
# pydub>=0.25.1

# Web automation (for ChatGPT fallback - only if needed)
# Uncomment if you want ChatGPT web automation fallback:
# selenium>=4.15.0
# webdriver-manager>=4.0.1

# System utilities
schedule>=1.2.0

# Development and testing (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0